FS-3664: allow viewing submitted application tasklists
This change allows the `/tasklist` route to load for funds that have
been closed and applications that have been otherwise submitted.
Additionally we check to see if those conditions are true and pass a new
flag `is_read_only_summary` to the form runner which will present that
tasklist sections summary screen without the option to modify your
answers.

This should allow applicants to see answers they've previously submitted
for complete submissions as well as for rounds in the past.