FS-3664: use existing time format utils
Favour using the existing string to datetime utility for displaying the
application submitted date. Note that an additional utility method is
added as the application submitted time seems to be recorded to
microsecond precision.

This works for the time being but it would be good for the date format
returned from stores to be in a consistent and guaranteed format.